requirements.txt
discord.py
openai~=0.27.4
PyYAML~=6.0
diffusers~=0.16.1
torch~=2.0.0
transformers~=4.11.3
requests~=2.28.2
accelerate